+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 ### Fixed
 
+- Split duration among partitions correctly [#336](https://github.com/Mapotempo/optimizer-api/pull/336)
 - Fix find_best_heuristic selection logic [#337](https://github.com/Mapotempo/optimizer-api/pull/337)
 
 ## [v1.8.2] - 2022-01-19